+++ src/NotFound.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const NotFound = () => {
+  return (
+    <div>
+      <h2>Page Not Found</h2>
+      <p>Sorry, the page you are looking for does not exist.</p>
+    </div>
+  );
+};
+
+export default NotFound;
