+++ src/components/PopularCities/PopularCities.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Row, Col } from 'react-bootstrap';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import ZymoOtherCities from '../../js/ZymoOtherCities';
 
 const PopularCities = ({ onCitySelect }) => {
   const [UserLocation, setUserLocation] = useState({ latitude: null, longitude: null });
@@ -34,53 +35,53 @@ const PopularCities = ({ onCitySelect }) => {
     { name: 'Vadodara' },
   ];
 
-  const ZymoOtherCities = [
-    { name: 'Bangalore' },
-    { name: 'Hyderabad' },
-    { name: 'Mumbai' },
-    { name: 'Delhi' },
-    { name: 'Chennai' },
-    { name: 'Pune' },
-    { name: 'Mangalore' },
-    { name: 'Dombivli' },
-    { name: 'Palava' },
-    { name: 'Thane' },
-    { name: 'Amritsar' },
-    { name: 'Kolkata' },
-    { name: 'Ahmedabad' },
-    { name: 'Bhubaneswar' },
-    { name: 'Chandigarh' },
-    { name: 'Coimbatore' },
-    { name: 'Jaipur' },
-    { name: 'Kochi' },
-    { name: 'Nashik' },
-    { name: 'Madurai' },
-    { name: 'Ghaziabad' },
-    { name: 'Merrut' },
-    { name: 'Goa' },
-    { name: 'Lucknow' },
-    { name: 'Bhopal' },
-    { name: 'Guwahati' },
-    { name: 'Indore' },
-    { name: 'Cochin' },
-    { name: 'Mysore' },
-    { name: 'Modinagar' },
-    { name: 'Muradnagar' },
-    { name: 'Gurugram' },
-    { name: 'Noida' },
-    { name: 'Nagpur' },
-    { name: 'Siliguri' },
-    { name: 'Trichy' },
-    { name: 'Vadodara' },
-    { name: 'Vijaywada' },
-    { name: 'Vijag' },
-    { name: 'Udupi' },
-    { name: 'Vishakapatnam' },
-    { name: 'Udaipur' },
-    { name: 'Jodhpur' },
-    { name: 'Haridwar' },
-    { name: 'Rishikesh' }
-  ];
+  // const ZymoOtherCities = [
+  //   { name: 'Bangalore' },
+  //   { name: 'Hyderabad' },
+  //   { name: 'Mumbai' },
+  //   { name: 'Delhi' },
+  //   { name: 'Chennai' },
+  //   { name: 'Pune' },
+  //   { name: 'Mangalore' },
+  //   { name: 'Dombivli' },
+  //   { name: 'Palava' },
+  //   { name: 'Thane' },
+  //   { name: 'Amritsar' },
+  //   { name: 'Kolkata' },
+  //   { name: 'Ahmedabad' },
+  //   { name: 'Bhubaneswar' },
+  //   { name: 'Chandigarh' },
+  //   { name: 'Coimbatore' },
+  //   { name: 'Jaipur' },
+  //   { name: 'Kochi' },
+  //   { name: 'Nashik' },
+  //   { name: 'Madurai' },
+  //   { name: 'Ghaziabad' },
+  //   { name: 'Merrut' },
+  //   { name: 'Goa' },
+  //   { name: 'Lucknow' },
+  //   { name: 'Bhopal' },
+  //   { name: 'Guwahati' },
+  //   { name: 'Indore' },
+  //   { name: 'Cochin' },
+  //   { name: 'Mysore' },
+  //   { name: 'Modinagar' },
+  //   { name: 'Muradnagar' },
+  //   { name: 'Gurugram' },
+  //   { name: 'Noida' },
+  //   { name: 'Nagpur' },
+  //   { name: 'Siliguri' },
+  //   { name: 'Trichy' },
+  //   { name: 'Vadodara' },
+  //   { name: 'Vijaywada' },
+  //   { name: 'Vijag' },
+  //   { name: 'Udupi' },
+  //   { name: 'Vishakapatnam' },
+  //   { name: 'Udaipur' },
+  //   { name: 'Jodhpur' },
+  //   { name: 'Haridwar' },
+  //   { name: 'Rishikesh' }
+  // ];
   
 
   useEffect(() => {
