+++ src/js/ZymoOtherCities.js
@@ -1,0 +1,50 @@
+const ZymoOtherCities = [
+    { name: 'Bangalore' },
+    { name: 'Hyderabad' },
+    { name: 'Mumbai' },
+    { name: 'Delhi' },
+    { name: 'Chennai' },
+    { name: 'Pune' },
+    { name: 'Mangalore' },
+    { name: 'Dombivli' },
+    { name: 'Palava' },
+    { name: 'Thane' },
+    { name: 'Amritsar' },
+    { name: 'Kolkata' },
+    { name: 'Ahmedabad' },
+    { name: 'Bhubaneswar' },
+    { name: 'Chandigarh' },
+    { name: 'Coimbatore' },
+    { name: 'Jaipur' },
+    { name: 'Kochi' },
+    { name: 'Nashik' },
+    { name: 'Madurai' },
+    { name: 'Ghaziabad' },
+    { name: 'Merrut' },
+    { name: 'Goa' },
+    { name: 'Lucknow' },
+    { name: 'Bhopal' },
+    { name: 'Guwahati' },
+    { name: 'Indore' },
+    { name: 'Cochin' },
+    { name: 'Mysore' },
+    { name: 'Modinagar' },
+    { name: 'Muradnagar' },
+    { name: 'Gurugram' },
+    { name: 'Noida' },
+    { name: 'Nagpur' },
+    { name: 'Siliguri' },
+    { name: 'Trichy' },
+    { name: 'Vadodara' },
+    { name: 'Vijaywada' },
+    { name: 'Vijag' },
+    { name: 'Udupi' },
+    { name: 'Vishakapatnam' },
+    { name: 'Udaipur' },
+    { name: 'Jodhpur' },
+    { name: 'Haridwar' },
+    { name: 'Rishikesh' }
+  ];
+  
+  export default ZymoOtherCities;
+
